feat: implement websocket telemetry for RPC and harden dev skill
