+++ aura-daemon/Cargo.toml
@@ -5,7 +5,7 @@ edition = "2021"
 
 [dependencies]
 aura-core = { path = "../aura-core" }
-axum = "0.7"
+axum = { version = "0.7", features = ["ws"] }
 tokio = { version = "1", features = ["full"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
@@ -13,3 +13,7 @@ tracing = "0.1"
 tracing-subscriber = "0.3"
 rand = { version = "0.10", features = ["std", "std_rng", "thread_rng"] }
 tower-http = { version = "0.6", features = ["cors"] }
+futures-util = "0.3"
+
+[dev-dependencies]
+tokio-tungstenite = "0.24"
